+++ packages/ops-access/ssh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elinpf/dsh-ops-access-ssh
 
+## 0.1.7
+
+### Patch Changes
+
+- @elinpf/dsh-ops-access@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes