+++ packages/ops-tool-environment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @elinpf/dsh-ops-tool-environment
 
+## 0.1.7
+
+### Patch Changes
+
+- @elinpf/dsh-ops-access@0.1.7
+  - @elinpf/dsh-ops-prompts@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes